requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-jose
pika
requests
